bikesharing/requirements.txt
# requirements.txt
# numpy==1.24.3
# pandas==2.0.3
# scikit-learn==1.3.0
# seaborn==0.12.2
# matplotlib==3.7.2
# pyyaml==6.0.1
# python-dateutil==2.8.2
# joblib==1.3.1
numpy
pandas
scikit-learn
seaborn
matplotlib
pyyaml
python-dateutil
joblib